Improved 'mise clean command' and setup guide

+++ s01-project-setup/cpp-project/.mise-tasks/clean.bash
@@ -3,4 +3,10 @@
 #MISE description="🧼 Delete the 'build' directory"
 #MISE quiet=true
 
+if [ ! -d build ]; then
+    printf "\n%s\n\n" '✅ No build directory found'
+    exit 0
+fi
+
 rm -rf build
+printf "\n%s\n\n" '✅ The build directory has been deleted'
